extract game selectors to separate file
