+++ src/modules/Game/gameReducer.test.ts
@@ -196,59 +196,3 @@ describe("Game reducer", () => {
 		expect(state === initState).toBeTruthy();
 	});
 });
-
-describe("Game gameSelectors ", () => {
-	it("gameIsPlaying", () => {
-		expect(gameSelectors.gameIsPlaying(initialAppState)).toBeFalsy();
-		expect(gameSelectors.gameIsPlaying({
-			...initialAppState,
-			game: {
-				...initGameState,
-				isPlaying: true
-			}
-		})).toBeTruthy();
-	});
-	it("hasPrevStep", () => {
-		expect(gameSelectors.hasPrevStep(initialAppState)).toBeFalsy();
-		expect(gameSelectors.hasPrevStep({
-			...initialAppState,
-			game: {
-				...initGameState,
-				isPlaying: true
-			}
-		})).toBeFalsy();
-		expect(gameSelectors.hasPrevStep({
-			...initialAppState,
-			game: {
-				...initGameState,
-				currentHistoryStep: 1
-			}
-		})).toBeTruthy();
-	});
-
-	it("hasNextStep", () => {
-		expect(gameSelectors.hasNextStep(initialAppState)).toBeFalsy();
-		expect(gameSelectors.hasNextStep({
-			...initialAppState,
-			game: {
-				...initGameState,
-				isPlaying: true
-			}
-		})).toBeFalsy();
-		expect(gameSelectors.hasNextStep({
-			...initialAppState,
-			game: {
-				...initGameState,
-				currentHistoryStep: 1
-			}
-		})).toBeFalsy();
-		expect(gameSelectors.hasNextStep({
-			...initialAppState,
-			game: {
-				...initGameState,
-				history: [[], [], []],
-				currentHistoryStep: 1
-			}
-		})).toBeTruthy();
-	});
-});
